Updating BoundaryCondition tests
